Fix navigation bar style issues
